Refactored to move the code to run as a different user into seperate files. Exposed the Windows/Linux differences in an adapter.
