+++ source/SilentProcessRunner/SilentProcessRunner.cs
@@ -14,14 +14,16 @@ using System.Text;
 using System.Threading;
 using Microsoft.Win32.SafeHandles;
 using Octopus.Diagnostics;
+using Octopus.SilentProcessRunner.Nix;
+using Octopus.SilentProcessRunner.Windows;
 
 namespace Octopus.SilentProcessRunner
 {
     public static class SilentProcessRunner
     {
-        static readonly object EnvironmentVariablesCacheLock = new object();
-        static IDictionary mostRecentMachineEnvironmentVariables = Environment.GetEnvironmentVariables(EnvironmentVariableTarget.Machine);
-        static readonly Dictionary<string, Dictionary<string, string>> EnvironmentVariablesForUserCache = new Dictionary<string, Dictionary<string, string>>();
+        static readonly IXPlatAdapter XPlatAdapter = RuntimeInformation.IsOSPlatform(OSPlatform.Windows)
+            ? (IXPlatAdapter)new WindowsAdapter()
+            : new NixAdapter();
 
         public static int ExecuteCommand(
             string executable,
@@ -124,10 +126,8 @@ namespace Octopus.SilentProcessRunner
 
                     if (runAs == null)
                         RunAsSameUser(process.StartInfo, customEnvironmentVariables);
-                    else if (PlatformDetection.IsRunningOnWindows)
-                        RunAsDifferentUser(process.StartInfo, runAs, customEnvironmentVariables);
                     else
-                        throw new PlatformNotSupportedException("NetCore on Linux or Mac does not support running a process as a different user.");
+                        XPlatAdapter.RunAsDifferentUser(process.StartInfo, runAs, customEnvironmentVariables);
 
                     process.StartInfo.RedirectStandardOutput = true;
                     process.StartInfo.RedirectStandardError = true;
@@ -242,10 +242,8 @@ namespace Octopus.SilentProcessRunner
 
                     if (runAs == null)
                         RunAsSameUser(process.StartInfo, customEnvironmentVariables);
-                    else if (PlatformDetection.IsRunningOnWindows)
-                        RunAsDifferentUser(process.StartInfo, runAs, customEnvironmentVariables);
                     else
-                        throw new PlatformNotSupportedException("NetCore on linux or Mac does not support running a process as a different user.");
+                        XPlatAdapter.RunAsDifferentUser(process.StartInfo, runAs, customEnvironmentVariables);
 
                     process.Start();
                 }
@@ -263,96 +261,6 @@ namespace Octopus.SilentProcessRunner
             if (customEnvironmentVariables != null && customEnvironmentVariables.Any())
                 foreach (var variable in customEnvironmentVariables)
                     processStartInfo.EnvironmentVariables[variable.Key] = variable.Value;
-        }
-
-        static void RunAsDifferentUser(ProcessStartInfo startInfo, NetworkCredential runAs, IDictionary<string, string>? customEnvironmentVariables)
-        {
-#pragma warning disable PC001 // API not supported on all platforms
-            startInfo.UserName = runAs.UserName;
-            startInfo.Domain = runAs.Domain;
-            startInfo.Password = runAs.SecurePassword;
-            startInfo.LoadUserProfile = true;
-#pragma warning restore PC001 // API not supported on all platforms
-
-            WindowStationAndDesktopAccess.GrantAccessToWindowStationAndDesktop(runAs.UserName, runAs.Domain);
-
-            if (customEnvironmentVariables != null && customEnvironmentVariables.Any())
-                SetEnvironmentVariablesForTargetUser(startInfo, runAs, customEnvironmentVariables);
-        }
-
-        static void SetEnvironmentVariablesForTargetUser(ProcessStartInfo startInfo, NetworkCredential runAs, IDictionary<string, string> customEnvironmentVariables)
-        {
-            // Double check before we go doing p/invoke gymnastics
-            if (customEnvironmentVariables == null || !customEnvironmentVariables.Any()) return;
-
-            // If ProcessStartInfo.enviromentVariables (field) is null, the new process will build its environment variables from scratch
-            // This will be the system environment variables, plus the user's profile variables (if the user profile is loaded)
-            // However, if the ProcessStartInfo.environmentVariables (field) is not null, these environment variables will be used instead
-            // As soon as we touch ProcessStartInfo.EnvironmentVariables (property) it lazy loads the environment variables for the current process
-            // which in turn means the launched process will get the environment variables for the wrong user profile!
-
-            // See https://msdn.microsoft.com/en-us/library/windows/desktop/ms682425(v=vs.85).aspx (CreateProcess) used when ProcessStartInfo.Username is not set
-            // See https://msdn.microsoft.com/en-us/library/windows/desktop/ms682431(v=vs.85).aspx (CreateProcessWithLogonW) used when ProcessStartInfo.Username is set
-
-            // Start by getting the environment variables for the target user (as if they started a process themselves)
-            // This will get the system environment variables along with the user's profile variables
-            var targetUserEnvironmentVariables = GetTargetUserEnvironmentVariables(runAs);
-
-            // Now copy in the extra environment variables we want to propagate from this process
-            foreach (var variable in customEnvironmentVariables)
-                targetUserEnvironmentVariables[variable.Key] = variable.Value;
-
-            // Starting from a clean slate, copy the resulting environment variables into the ProcessStartInfo
-            startInfo.EnvironmentVariables.Clear();
-            foreach (var variable in targetUserEnvironmentVariables)
-                startInfo.EnvironmentVariables[variable.Key] = variable.Value;
-        }
-
-        static Dictionary<string, string> GetTargetUserEnvironmentVariables(NetworkCredential runAs)
-        {
-            var cacheKey = $"{runAs.Domain}\\{runAs.UserName}";
-
-            // Start with a pessimistic lock until such a time where we want more throughput for concurrent processes
-            // In the real world we shouldn't have too many processes wanting to run concurrently
-            lock (EnvironmentVariablesCacheLock)
-            {
-                // If the machine environment variables have changed we should invalidate the entire cache
-                InvalidateEnvironmentVariablesForUserCacheIfMachineEnvironmentVariablesHaveChanged();
-
-                // Otherwise the cache will generally be valid, except for the (hopefully) rare case where a variable was added/changed for the specific user
-                if (EnvironmentVariablesForUserCache.TryGetValue(cacheKey, out var cached))
-                    return cached;
-
-                Dictionary<string, string> targetUserEnvironmentVariables;
-                using (var token = AccessToken.Logon(runAs.UserName, runAs.Password, runAs.Domain))
-                using (var userProfile = UserProfile.Load(token))
-                {
-                    targetUserEnvironmentVariables = EnvironmentBlock.GetEnvironmentVariablesForUser(token, false);
-                    userProfile.Unload();
-                }
-
-                // Cache the target user's environment variables so we don't have to load them every time
-                // The downside is that once we target a certain user account, their variables are snapshotted in time
-                EnvironmentVariablesForUserCache[cacheKey] = targetUserEnvironmentVariables;
-                return targetUserEnvironmentVariables;
-            }
-        }
-
-        static void InvalidateEnvironmentVariablesForUserCacheIfMachineEnvironmentVariablesHaveChanged()
-        {
-            var currentMachineEnvironmentVariables = Environment.GetEnvironmentVariables(EnvironmentVariableTarget.Machine);
-            var machineEnvironmentVariablesHaveChanged =
-#pragma warning disable DE0006 // API is deprecated
-                !currentMachineEnvironmentVariables.Cast<DictionaryEntry>()
-                    .OrderBy(e => e.Key)
-                    .SequenceEqual(mostRecentMachineEnvironmentVariables.Cast<DictionaryEntry>().OrderBy(e => e.Key));
-#pragma warning restore DE0006 // API is deprecated
-
-            if (machineEnvironmentVariablesHaveChanged)
-            {
-                mostRecentMachineEnvironmentVariables = currentMachineEnvironmentVariables;
-                EnvironmentVariablesForUserCache.Clear();
-            }
         }
 
         static void DoOurBestToCleanUp(System.Diagnostics.Process process, Action<string> error)
@@ -490,405 +398,9 @@ namespace Octopus.SilentProcessRunner
             }
         }
 
-        // Required to allow a service to run a process as another user
-        // See http://stackoverflow.com/questions/677874/starting-a-process-with-credentials-from-a-windows-service/30687230#30687230
-        internal class WindowStationAndDesktopAccess
-        {
-            public static void GrantAccessToWindowStationAndDesktop(string username, string? domainName = null)
-            {
-                var hWindowStation = Win32Helper.Invoke(() => GetProcessWindowStation(), "Failed to get a handle to the current window station for this process");
-                const int windowStationAllAccess = 0x000f037f;
-                GrantAccess(username, domainName, hWindowStation, windowStationAllAccess);
-                var hDesktop = Win32Helper.Invoke(() => GetThreadDesktop(GetCurrentThreadId()), "Failed to the a handle to the desktop for the current thread");
-                const int desktopRightsAllAccess = 0x000f01ff;
-                GrantAccess(username, domainName, hDesktop, desktopRightsAllAccess);
-            }
 
-            static void GrantAccess(string username, string? domainName, IntPtr handle, int accessMask)
-            {
-                SafeHandle safeHandle = new NoopSafeHandle(handle);
-                var security =
-                    new GenericSecurity(
-                        false,
-                        ResourceType.WindowObject,
-                        safeHandle,
-                        AccessControlSections.Access);
 
-                var account = string.IsNullOrEmpty(domainName)
-#pragma warning disable PC001 // API not supported on all platforms
-                    ? new NTAccount(username)
-                    : new NTAccount(domainName, username);
-#pragma warning restore PC001 // API not supported on all platforms
 
-                security.AddAccessRule(
-                    new GenericAccessRule(
-                        account,
-                        accessMask,
-                        AccessControlType.Allow));
-                security.Persist(safeHandle, AccessControlSections.Access);
-            }
-
-            // Native API not available in UWP
-            // All the code to manipulate a security object is available in .NET framework,
-            // but its API tries to be type-safe and handle-safe, enforcing a special implementation
-            // (to an otherwise generic WinAPI) for each handle type. This is to make sure
-            // only a correct set of permissions can be set for corresponding object types and
-            // mainly that handles do not leak.
-            // Hence the AccessRule and the NativeObjectSecurity classes are abstract.
-            // This is the simplest possible implementation that yet allows us to make use
-            // of the existing .NET implementation, sparing necessity to
-            // P/Invoke the underlying WinAPI.
-            class GenericAccessRule : AccessRule
-            {
-                public GenericAccessRule(
-                    IdentityReference identity,
-                    int accessMask,
-                    AccessControlType type) :
-                    base(identity,
-                        accessMask,
-                        false,
-                        InheritanceFlags.None,
-                        PropagationFlags.None,
-                        type)
-                {
-                }
-            }
-
-            class GenericSecurity : NativeObjectSecurity
-            {
-                public GenericSecurity(
-                    bool isContainer,
-                    ResourceType resType,
-                    SafeHandle objectHandle,
-                    AccessControlSections sectionsRequested)
-                    : base(isContainer, resType, objectHandle, sectionsRequested)
-                {
-                }
-
-                public override Type AccessRightType => throw new NotImplementedException();
-
-                public override Type AccessRuleType => typeof(AccessRule);
-
-                public override Type AuditRuleType => typeof(AuditRule);
-
-                public new void Persist(SafeHandle handle, AccessControlSections includeSections)
-                {
-#pragma warning disable PC001 // API not supported on all platforms
-                    base.Persist(handle, includeSections);
-#pragma warning restore PC001 // API not supported on all platforms
-                }
-
-                public new void AddAccessRule(AccessRule rule)
-                {
-#pragma warning disable PC001 // API not supported on all platforms
-                    base.AddAccessRule(rule);
-#pragma warning restore PC001 // API not supported on all platforms
-                }
-
-                public override AccessRule AccessRuleFactory(
-                    IdentityReference identityReference,
-                    int accessMask,
-                    bool isInherited,
-                    InheritanceFlags inheritanceFlags,
-                    PropagationFlags propagationFlags,
-                    AccessControlType type)
-                    => throw new NotImplementedException();
-
-                public override AuditRule AuditRuleFactory(
-                    IdentityReference identityReference,
-                    int accessMask,
-                    bool isInherited,
-                    InheritanceFlags inheritanceFlags,
-                    PropagationFlags propagationFlags,
-                    AuditFlags flags)
-                    => throw new NotImplementedException();
-            }
-
-            // Handles returned by GetProcessWindowStation and GetThreadDesktop should not be closed
-            class NoopSafeHandle : SafeHandle
-            {
-                public NoopSafeHandle(IntPtr handle) :
-                    base(handle, false)
-                {
-                }
-
-                public override bool IsInvalid => false;
-
-                protected override bool ReleaseHandle()
-                    => true;
-            }
-
-#pragma warning disable PC003 // Native API not available in UWP
-            [DllImport("user32.dll", SetLastError = true)]
-            static extern IntPtr GetProcessWindowStation();
-
-            [DllImport("user32.dll", SetLastError = true)]
-            static extern IntPtr GetThreadDesktop(int dwThreadId);
-
-            [DllImport("kernel32.dll", SetLastError = true)]
-            static extern int GetCurrentThreadId();
-#pragma warning restore PC003
-        }
-
-        internal class AccessToken : IDisposable
-        {
-            public enum LogonProvider
-            {
-                Default = 0,
-                WinNT40 = 2,
-                WinNT50 = 3
-            }
-
-            public enum LogonType
-            {
-                Interactive = 2,
-                Network = 3,
-                Batch = 4,
-                Service = 5,
-                Unlock = 7,
-                NetworkClearText = 8,
-                NewCredentials = 9
-            }
-
-            AccessToken(string username, SafeAccessTokenHandle handle)
-            {
-                Username = username;
-                Handle = handle;
-            }
-
-            public string Username { get; }
-            public SafeAccessTokenHandle Handle { get; }
-
-            public static AccessToken Logon(string username,
-                string password,
-                string domain = ".",
-                LogonType logonType = LogonType.Network,
-                LogonProvider logonProvider = LogonProvider.Default)
-            {
-                // See https://msdn.microsoft.com/en-us/library/windows/desktop/aa378184(v=vs.85).aspx
-                var hToken = IntPtr.Zero;
-                Win32Helper.Invoke(() => LogonUser(username,
-                        domain,
-                        password,
-                        LogonType.Network,
-                        LogonProvider.Default,
-                        out hToken),
-                    $"Logon failed for the user '{username}'");
-
-                return new AccessToken(username, new SafeAccessTokenHandle(hToken));
-            }
-
-            public void Dispose()
-            {
-                Handle?.Dispose();
-            }
-
-            [DllImport("advapi32.dll", SetLastError = true)]
-#pragma warning disable PC003 // Native API not available in UWP
-            static extern bool LogonUser(string username,
-                string domain,
-                string password,
-                LogonType logonType,
-                LogonProvider logonProvider,
-                out IntPtr hToken);
-#pragma warning restore PC003 // Native API not available in UWP
-        }
-
-        internal class UserProfile : IDisposable
-        {
-            readonly AccessToken token;
-            readonly SafeRegistryHandle userProfile;
-
-            UserProfile(AccessToken token, SafeRegistryHandle userProfile)
-            {
-                this.token = token;
-                this.userProfile = userProfile;
-            }
-
-            public static UserProfile Load(AccessToken token)
-            {
-                var userProfile = new PROFILEINFO
-                {
-                    lpUserName = token.Username
-                };
-                userProfile.dwSize = Marshal.SizeOf(userProfile);
-
-                // See https://msdn.microsoft.com/en-us/library/windows/desktop/bb762281(v=vs.85).aspx
-                Win32Helper.Invoke(() => LoadUserProfile(token.Handle, ref userProfile),
-                    $"Failed to load user profile for user '{token.Username}'");
-
-                return new UserProfile(token, new SafeRegistryHandle(userProfile.hProfile, false));
-            }
-
-            public void Unload()
-            {
-                // See https://msdn.microsoft.com/en-us/library/windows/desktop/bb762282(v=vs.85).aspx
-                // This function closes the registry handle for the user profile too
-                Win32Helper.Invoke(() => UnloadUserProfile(token.Handle, userProfile),
-                    $"Failed to unload user profile for user '{token.Username}'");
-            }
-
-            public void Dispose()
-            {
-                if (userProfile != null && !userProfile.IsClosed)
-                {
-                    try
-                    {
-                        Unload();
-                    }
-                    catch
-                    {
-                        // Don't throw in dispose method
-                    }
-
-                    userProfile.Dispose();
-                }
-            }
-
-            [StructLayout(LayoutKind.Sequential)]
-            struct PROFILEINFO
-            {
-                public int dwSize;
-                public readonly int dwFlags;
-                public string lpUserName;
-                public readonly string lpProfilePath;
-                public readonly string lpDefaultPath;
-                public readonly string lpServerName;
-                public readonly string lpPolicyPath;
-                public readonly IntPtr hProfile;
-            }
-
-#pragma warning disable PC003 // Native API not available in UWP
-            [DllImport("userenv.dll", SetLastError = true)]
-            static extern bool LoadUserProfile(SafeAccessTokenHandle hToken, ref PROFILEINFO lpProfileInfo);
-
-            [DllImport("userenv.dll", SetLastError = true)]
-            static extern bool UnloadUserProfile(SafeAccessTokenHandle hToken, SafeRegistryHandle hProfile);
-#pragma warning restore PC003 // Native API not available in UWP
-        }
-
-        class Win32Helper
-        {
-            public static bool Invoke(Func<bool> nativeMethod, string failureDescription)
-            {
-                try
-                {
-                    return nativeMethod() ? true : throw new Win32Exception();
-                }
-                catch (Win32Exception ex)
-                {
-                    throw new Exception($"{failureDescription}: {ex.Message}", ex);
-                }
-            }
-
-            public static T Invoke<T>(Func<T> nativeMethod, Func<T, bool> successful, string failureDescription)
-            {
-                try
-                {
-                    var result = nativeMethod();
-                    return successful(result) ? result : throw new Win32Exception();
-                }
-                catch (Win32Exception ex)
-                {
-                    throw new Exception($"{failureDescription}: {ex.Message}", ex);
-                }
-            }
-
-            public static IntPtr Invoke(Func<IntPtr> nativeMethod, string failureDescription)
-            {
-                try
-                {
-                    var result = nativeMethod();
-                    return result != IntPtr.Zero ? result : throw new Win32Exception();
-                }
-                catch (Win32Exception ex)
-                {
-                    throw new Exception($"{failureDescription}: {ex.Message}", ex);
-                }
-            }
-        }
-
-        internal class EnvironmentBlock
-        {
-            internal static Dictionary<string, string> GetEnvironmentVariablesForUser(AccessToken token, bool inheritFromCurrentProcess)
-            {
-                var env = IntPtr.Zero;
-
-                // See https://msdn.microsoft.com/en-us/library/windows/desktop/bb762270(v=vs.85).aspx
-                Win32Helper.Invoke(() => CreateEnvironmentBlock(out env, token.Handle, inheritFromCurrentProcess),
-                    $"Failed to load the environment variables for the user '{token.Username}'");
-
-                var userEnvironment = new Dictionary<string, string>();
-                try
-                {
-                    var testData = new StringBuilder();
-                    unsafe
-                    {
-                        // The environment block is an array of null-terminated Unicode strings.
-                        // Key and Value are separated by =
-                        // The list ends with two nulls (\0\0).
-                        var start = (short*)env.ToPointer();
-                        var done = false;
-                        var current = start;
-                        while (!done)
-                        {
-                            if (testData.Length > 0 && *current == 0 && current != start)
-                            {
-                                var data = testData.ToString();
-                                var index = data.IndexOf('=');
-                                if (index == -1)
-                                    userEnvironment.Add(data, "");
-                                else if (index == data.Length - 1)
-                                    userEnvironment.Add(data.Substring(0, index), "");
-                                else
-                                    userEnvironment.Add(data.Substring(0, index), data.Substring(index + 1));
-                                testData.Length = 0;
-                            }
-
-                            if (*current == 0 && current != start && *(current - 1) == 0)
-                                done = true;
-                            if (*current != 0)
-                                testData.Append((char)*current);
-                            current++;
-                        }
-                    }
-                }
-                finally
-                {
-                    // See https://msdn.microsoft.com/en-us/library/windows/desktop/bb762274(v=vs.85).aspx
-                    Win32Helper.Invoke(() => DestroyEnvironmentBlock(env),
-                        $"Failed to destroy the environment variables structure for user '{token.Username}'");
-                }
-
-                return userEnvironment;
-            }
-
-#pragma warning disable PC003 // Native API not available in UWP
-            [DllImport("userenv.dll", SetLastError = true)]
-            static extern bool CreateEnvironmentBlock(out IntPtr lpEnvironment, SafeAccessTokenHandle hToken, bool inheritFromCurrentProcess);
-
-            [DllImport("userenv.dll", SetLastError = true)]
-            static extern bool DestroyEnvironmentBlock(IntPtr lpEnvironment);
-#pragma warning restore PC003 // Native API not available in UWP
-        }
-
-        internal sealed class SafeAccessTokenHandle : SafeHandle
-        {
-            // 0 is an Invalid Handle
-            public SafeAccessTokenHandle(IntPtr handle) : base(handle, true)
-            {
-            }
-
-            public static SafeAccessTokenHandle InvalidHandle => new SafeAccessTokenHandle(IntPtr.Zero);
-
-            public override bool IsInvalid => handle == IntPtr.Zero || handle == new IntPtr(-1);
-
-            protected override bool ReleaseHandle()
-                => CloseHandle(handle);
-
-            [DllImport("kernel32.dll", SetLastError = true)]
-            static extern bool CloseHandle(IntPtr hHandle);
-        }
 
         // ReSharper disable InconsistentNaming
         const int MAX_DEFAULTCHAR = 2;
